Rename `defaultMemoize.ts` to `lruMemoize.ts`
